+++ server/src/internal/balances/track/v3/runTrackV3.ts
@@ -13,10 +13,8 @@ import type { AutumnContext } from "@/honoUtils/HonoEnv.js";
 import { getOrCreateCachedFullSubject } from "@/internal/customers/cache/fullSubject/actions/getOrCreateCachedFullSubject.js";
 import { getOrSetCachedFullSubject } from "@/internal/customers/cache/fullSubject/actions/getOrSetCachedFullSubject.js";
 import type { FeatureDeduction } from "../../utils/types/featureDeduction.js";
-import {
-	getTrackIdempotencyKey,
-} from "../utils/handleEventIdempotencyKey.js";
 import { runRedisTrackV3 } from "./runRedisTrackV3.js";
+import { getTrackIdempotencyKey } from "./trackIdempotencyKey.js";
 
 const getTrackFullSubject = async ({
 	ctx,
